refactor: main에 사용되는 함수들을 유형에 맞추어 분리(/urils)
answer.js : 게임 결과에 관련한 함수들
score.js : 게임 결과 관련
userInput.js : 사용자 입력 관련
util.js : 전역적으로 사용 가능한 유틸 함수들
